chore: correcting lint & code style - move to typesafe config - move from code first -> spec first - add oapi-codegen with gin0server support - delete decimal type

+++ laboratory_work-6/shared/responses/status.go
@@ -1,6 +1,6 @@
 package responses
 
 type StatusResponse struct {
-	Code    uint16 `json:"code"`
+	Code    int    `json:"code"`
 	Message string `json:"message"`
 }
