+++ laboratory_work-6/shared/responses/success.go
@@ -1,7 +1,7 @@
 package responses
 
 type SuccessResponse[T any] struct {
-	Code    uint16 `json:"code"`
+	Code    int    `json:"code"`
 	Message string `json:"message"`
 	Data    T      `json:"data"`
 }
